+++ src/ir/CategoricalDistr.h
@@ -16,10 +16,10 @@ public:
 
   size_t size() const;
   void addWeight(double w);
-  double getWeight(int k) const;
+  double getWeight(size_t k) const;
   const std::vector<double>& getWeights() const;
   void addSymbol(std::shared_ptr<ConstSymbol> c);
-  std::shared_ptr<ConstSymbol> getSymbol(int k) const;
+  std::shared_ptr<ConstSymbol> getSymbol(size_t k) const;
   const std::vector<std::shared_ptr<ConstSymbol>>& getSymbols() const;
   void print(FILE* file, int indent) const;
